#348: Move actions in correct folder
