--- Frontend/ShoppingList.Frontend.Redux/Recipes/Actions/Editor/AddIngredientToShoppingListChangedAction.cs
@@ -1,3 +1,0 @@
-﻿namespace ProjectHermes.ShoppingList.Frontend.Redux.Recipes.Actions.Editor;
-
-public record AddIngredientToShoppingListChangedAction(Guid IngredientKey, bool AddToShoppingList);
