+++ Frontend/ShoppingList.Frontend.Redux/Recipes/Actions/Editor/AddToShoppingListModal/AddIngredientToShoppingListChangedAction.cs
@@ -1,0 +1,3 @@
+﻿namespace ProjectHermes.ShoppingList.Frontend.Redux.Recipes.Actions.Editor.AddToShoppingListModal;
+
+public record AddIngredientToShoppingListChangedAction(Guid IngredientKey, bool AddToShoppingList);
