+++ Frontend/ShoppingList.Frontend.Redux/Recipes/Actions/Editor/AddToShoppingListModal/AddToShoppingListNumberOfServingsChangedAction.cs
@@ -1,0 +1,3 @@
+﻿namespace ProjectHermes.ShoppingList.Frontend.Redux.Recipes.Actions.Editor.AddToShoppingListModal;
+
+public record AddToShoppingListNumberOfServingsChangedAction(int NumberOfServings);
