+++ Frontend/ShoppingList.Frontend.Redux/Recipes/Actions/Editor/AddToShoppingListModal/LoadAddToShoppingListAction.cs
@@ -1,0 +1,2 @@
+﻿namespace ProjectHermes.ShoppingList.Frontend.Redux.Recipes.Actions.Editor.AddToShoppingListModal;
+public record LoadAddToShoppingListAction;
