+++ Frontend/ShoppingList.Frontend.Redux/Recipes/Actions/Editor/AddToShoppingListModal/LoadAddToShoppingListStartedAction.cs
@@ -1,0 +1,2 @@
+﻿namespace ProjectHermes.ShoppingList.Frontend.Redux.Recipes.Actions.Editor.AddToShoppingListModal;
+public record LoadAddToShoppingListStartedAction;
