--- Frontend/ShoppingList.Frontend.Redux/Recipes/Actions/Editor/AddToShoppingListNumberOfServingsChangedAction.cs
@@ -1,3 +1,0 @@
-﻿namespace ProjectHermes.ShoppingList.Frontend.Redux.Recipes.Actions.Editor;
-
-public record AddToShoppingListNumberOfServingsChangedAction(int NumberOfServings);
