--- Frontend/ShoppingList.Frontend.Redux/Recipes/Actions/Editor/LoadAddToShoppingListAction.cs
@@ -1,2 +1,0 @@
-﻿namespace ProjectHermes.ShoppingList.Frontend.Redux.Recipes.Actions.Editor;
-public record LoadAddToShoppingListAction;
