--- Frontend/ShoppingList.Frontend.Redux/Recipes/Actions/Editor/LoadAddToShoppingListStartedAction.cs
@@ -1,2 +1,0 @@
-﻿namespace ProjectHermes.ShoppingList.Frontend.Redux.Recipes.Actions.Editor;
-public record LoadAddToShoppingListStartedAction;
